Update DeveloperGuide.md and add Timetable Class
Separate UI and logic of timetable command by adding a new Timetable
class
