+++ src/main/resources/view/TimetableInfoPage.fxml
@@ -6,7 +6,7 @@
 <ScrollPane stylesheets="Timetable.css" fx:id="scrollPane" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1">
 <AnchorPane stylesheets="Timetable.css" maxHeight="400.0" maxWidth="600.0">
    <children>
-      <GridPane fx:id="timetable" stylesheets="Timetable.css" layoutX="-1.0" layoutY="-1.0" maxHeight="400.0" maxWidth="600.0" style="-fx-background-color: white;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+      <GridPane fx:id="timetableShown" stylesheets="Timetable.css" layoutX="-1.0" layoutY="-1.0" maxHeight="400.0" maxWidth="600.0" style="-fx-background-color: white;" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
       </GridPane>
 </children>
 </AnchorPane>
